Remove dead and useless code
- Removed stale code not used anywhere
- Removed defunct replacement code from .NET Core 1.x that can now use the real methods in .NET Core 2.0 (e.g. DnsAPMExtensions)
- Removed stale compilation constants
- Removed Unix-specific MultipleSocketMultipleConnectAsync that's no longer necessary as we're able to reuse Sockets after failed connects
- Moved some Unix-specific code out of the Windows compilation
- Modified some perf-counter code to minimize its impact, since it's currently a nop
- Replaced the same repeated snippet of error handling code with a helper
- Removed unused namespace imports
- Removed some unnecessary null checks, for objects that can never be null based on always being assigned to new objects in the ctors
- Removed unnecessary tracing guards, where a guard was just used at all call sites
- Removed invalid asserts that were triggered by new tests
- Some other minor miscellaneous cleanup as I noticed things, like using the null-coalescing operator
