+++ src/Common/src/Interop/Windows/Winsock/Interop.WSAGetOverlappedResult.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using System.Runtime.InteropServices;
 using System.Net.Sockets;
