+++ src/Common/src/Interop/Windows/Winsock/Interop.WinsockAsync.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Microsoft.Win32.SafeHandles;
-
 using System;
 using System.Net.Sockets;
 using System.Runtime.InteropServices;
