+++ src/Common/src/Interop/Windows/Winsock/Interop.select.cs
@@ -4,7 +4,6 @@
 
 using System;
 using System.Runtime.InteropServices;
-using System.Net.Sockets;
 
 internal static partial class Interop
 {
