+++ src/Common/src/System/Net/ContextAwareResult.Windows.cs
@@ -2,10 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Diagnostics;
-using System.Security;
 using System.Security.Principal;
-using System.Threading;
 
 namespace System.Net
 {
