+++ src/Common/src/System/Net/ContextAwareResult.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Diagnostics;
-using System.Security;
 using System.Threading;
 
 namespace System.Net
