+++ src/Common/src/System/Net/DebugSafeHandleMinusOneIsInvalid.cs
@@ -4,12 +4,6 @@
 
 using Microsoft.Win32.SafeHandles;
 
-using System.Net.NetworkInformation;
-using System.Net.Sockets;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-using System.Threading;
-
 namespace System.Net
 {
 #if DEBUG
