+++ src/Common/src/System/Net/IPAddressParserStatics.cs
@@ -2,9 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Net.Sockets;
-using System.Text;
-
 namespace System.Net
 {
     internal static class IPAddressParserStatics
