+++ src/Common/src/System/Net/LazyAsyncResult.cs
@@ -4,7 +4,6 @@
 
 using System.Threading;
 using System.Diagnostics;
-using System.Collections;
 using System.Threading.Tasks;
 
 namespace System.Net
