+++ src/Common/src/System/Net/SocketAddressPal.Windows.cs
@@ -2,10 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Diagnostics;
 using System.Net.Sockets;
-using System.Text;
 
 namespace System.Net
 {
