+++ src/Common/src/System/Threading/Tasks/TaskToApm.cs
@@ -17,7 +17,6 @@
 //     }
 
 using System.Diagnostics;
-using System.IO;
 
 namespace System.Threading.Tasks
 {
