+++ src/System.Net.Sockets/src/System/Net/SocketPerfCounters.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Diagnostics;
 using System.Threading;
 
 namespace System.Net
@@ -18,41 +19,13 @@ namespace System.Net
     internal sealed class SocketPerfCounter
     {
         private static SocketPerfCounter s_instance;
-        private static object s_lockObject = new object();
 
-        public static SocketPerfCounter Instance
-        {
-            get
-            {
-                if (Volatile.Read(ref s_instance) == null)
-                {
-                    lock (s_lockObject)
-                    {
-                        if (Volatile.Read(ref s_instance) == null)
-                        {
-                            s_instance = new SocketPerfCounter();
-                        }
-                    }
-                }
-                return s_instance;
-            }
-        }
+        public static SocketPerfCounter Instance => LazyInitializer.EnsureInitialized(ref s_instance);
 
-        public bool Enabled
-        {
-            get
-            {
-                // TODO (#7833): Implement socket perf counters.
-                return false;
-            }
-        }
+        public bool Enabled => false; // TODO (#7833): Implement socket perf counters.
 
-        public void Increment(SocketPerfCounterName perfCounter)
-        {
-            // TODO (#7833): Implement socket perf counters.
-        }
-
-        public void Increment(SocketPerfCounterName perfCounter, long amount)
+        [Conditional("TODO7833")]
+        public void Increment(SocketPerfCounterName perfCounter, long amount = 1)
         {
             // TODO (#7833): Implement socket perf counters.
         }
