+++ src/System.Net.Sockets/src/System/Net/Sockets/ConnectOverlappedAsyncResult.cs
@@ -2,12 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Net;
-using System.Runtime.InteropServices;
-using System.Threading;
-using Microsoft.Win32;
-
 namespace System.Net.Sockets
 {
     // ConnectOverlappedAsyncResult - used to take care of storage for async Socket BeginConnect call.
