+++ src/System.Net.Sockets/src/System/Net/Sockets/DynamicWinsockMethods.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Security;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
@@ -111,7 +110,7 @@ namespace System.Net.Sockets
                 return (T)(object)_transmitPackets;
             }
 
-            System.Diagnostics.Debug.Assert(false, "Invalid type passed to DynamicWinsockMethods.GetDelegate");
+            Debug.Fail("Invalid type passed to DynamicWinsockMethods.GetDelegate");
             return null;
         }
