+++ src/System.Net.Sockets/src/System/Net/Sockets/IPPacketInformation.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Net;
-
 namespace System.Net.Sockets
 {
     public struct IPPacketInformation
