+++ src/System.Net.Sockets/src/System/Net/Sockets/LingerOption.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-
 namespace System.Net.Sockets
 {
     // Contains information for a socket's linger time, the amount of time it will
