+++ src/System.Net.Sockets/src/System/Net/Sockets/OverlappedAsyncResult.cs
@@ -2,13 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Net;
-using System.Runtime.InteropServices;
-using System.Threading;
-using Microsoft.Win32;
-using System.Collections.Generic;
-
 namespace System.Net.Sockets
 {
     // OverlappedAsyncResult
