+++ src/System.Net.Sockets/src/System/Net/Sockets/SocketFlags.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-
 namespace System.Net.Sockets
 {
     // Provides constant values for socket messages.
