+++ src/System.Net.Sockets/src/System/Net/Sockets/TCPListener.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Net;
 using System.Threading.Tasks;
 
 namespace System.Net.Sockets
@@ -134,11 +132,6 @@ namespace System.Net.Sockets
 
             if (NetEventSource.IsEnabled) NetEventSource.Enter(this);
 
-            if (_serverSocket == null)
-            {
-                throw new InvalidOperationException(SR.net_InvalidSocketHandle);
-            }
-
             // Already listening.
             if (_active)
             {
@@ -167,12 +160,7 @@ namespace System.Net.Sockets
         {
             if (NetEventSource.IsEnabled) NetEventSource.Enter(this);
 
-            if (_serverSocket != null)
-            {
-                _serverSocket.Dispose();
-                _serverSocket = null;
-            }
-
+            _serverSocket.Dispose();
             _active = false;
             _serverSocket = new Socket(_serverSocketEP.AddressFamily, SocketType.Stream, ProtocolType.Tcp);
