+++ src/System.Net.Sockets/src/System/Net/Sockets/TransmitFileAsyncResult.Windows.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System.IO;
-using System.Runtime.InteropServices;
 
 namespace System.Net.Sockets
 {
