+++ src/System.Net.Sockets/src/System/Net/Sockets/UDPClient.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Diagnostics;
 using System.Threading.Tasks;
 
 namespace System.Net.Sockets
@@ -125,7 +124,6 @@ namespace System.Net.Sockets
         {
             get
             {
-                Debug.Assert(_clientSocket != null);
                 return _clientSocket;
             }
             set
