+++ src/System.Net.Sockets/src/System/Net/Sockets/UDPClient.netstandard.cs
@@ -2,9 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Diagnostics;
-using System.Threading.Tasks;
-
 namespace System.Net.Sockets
 {
     public partial class UdpClient
